float16 vs 32 in gen_noise_image.py
The previous noise setup allowed us to set precision in the ASDF output but this broke the FITS output. I think this will get the behavior we want for ASDF without breaking FITS.

+++ src/romanimpreprocess/L1_to_L2/gen_noise_image.py
@@ -381,7 +381,8 @@ def generate_all_noise(config):
         af.write_to(f)
     if "FITSOUT" in config:
         if config["FITSOUT"]:
-            fitsout = fits.HDUList([fits.PrimaryHDU(noiseimage)])
+            fitsout = fits.HDUList([fits.PrimaryHDU(noiseimage.astype(np.float32))])
+            # FITS doesn't have float16 so we convert to float32
             fitsout.writeto(config["NOISE"]["OUT"][:-5] + "_asdf_to.fits", overwrite=True)
